requirements.txt
streamlit
pandas
numpy
pandasql
yfinance
plotly.express
